alpaka::concepts::impl::View use alpaka::concepts::Api
- required restructuring of the header files to avoid circular includes
